improve metrics collection 2
