+++ pkg/eventservice/event_service_test.go
@@ -68,9 +68,6 @@ func TestEventServiceBasic(t *testing.T) {
 
 	log.Info("start event service basic test")
 
-	// Set the max task time range to a very large value to avoid the test case being blocked.
-	maxTaskTimeRange = time.Duration(math.MaxInt64)
-
 	mockStore := newMockEventStore(100)
 	mockStore.Run(ctx)
